+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### v1.1.0
+
+1. PHP Mailer added
+2. Padding fixed on app container
+3. Password reset feature added
+
 ### v1.0.1
 
 1. Page 404 added